+++ unicode-data-core/lib/Unicode/Char/General.hs
@@ -26,8 +26,6 @@ module Unicode.Char.General
     , isSeparator
     , isSymbol
     , isWhiteSpace
-    , isLetter
-    , isSpace
     -- ** Re-export
     , isAscii
     , isLatin1
@@ -430,25 +428,6 @@ isSymbol c = case generalCategory c of
     ModifierSymbol -> True
     OtherSymbol    -> True
     _              -> False
-
--- | Returns 'True' for alphabetic Unicode characters (lower-case, upper-case
--- and title-case letters, plus letters of caseless scripts and modifiers
--- letters).
---
--- @since 0.1.0
-{-# INLINE isLetter #-}
-{-# DEPRECATED isLetter "Use isAlphabetic instead. Note that the behavior of this function does not match base:Data.Char.isLetter. See Unicode.Char.General.Compat for behavior compatible with base:Data.Char." #-}
-isLetter :: Char -> Bool
-isLetter = P.isAlphabetic
-
--- | Returns 'True' for any whitespace characters, and the control
--- characters @\\t@, @\\n@, @\\r@, @\\f@, @\\v@.
---
--- @since 0.1.0
-{-# INLINE isSpace #-}
-{-# DEPRECATED isSpace "Use isWhiteSpace instead. Note that the behavior of this function does not match base:Data.Char.isSpace. See Unicode.Char.General.Compat for behavior compatible with base:Data.Char." #-}
-isSpace :: Char -> Bool
-isSpace = P.isWhite_Space
 
 -------------------------------------------------------------------------------
 -- Korean Hangul