+++ unicode-data-names/Changelog.md
@@ -1,6 +1,12 @@
 # Changelog
 
-## 0.2.0 (September 2022)
+## 15.0.0 (February 2023)
+
+- New version scheme: `U.B.M`, where `U` is the Unicode standard major version
+  number, `B` marks a breaking change and `M` a non-breaking change per
+  [PVP](https://pvp.haskell.org/).
+
+## 0.2.0 (October 2022)
 
 - Update to [Unicode 15.0.0](https://www.unicode.org/versions/Unicode15.0.0/).